sistema_reportes/sistema_reportes/urls.py

from django.contrib import admin
from django.urls import path, include
from rest_framework import permissions
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, SpectacularSwaggerView
from reportes.views import frontend_view
from reportes.views import integrar_snifa

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('reportes.urls')),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
    path('api/swagger/', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger-ui'),
    path('api/redoc/', SpectacularRedocView.as_view(url_name='schema'), name='redoc'),
    path('frontend/', frontend_view, name='frontend'),
    path('integrar-snifa/', integrar_snifa, name='integrar_snifa')
]
