sistema_reportes/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
psycopg2-binary==2.9.10
djangorestframework-camel-case==1.4.2
drf-spectacular==0.27.1
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.40.6
redis==5.0.1
celery==5.3.6
django-redis==5.3.0
